Element ScrollTo center
There are instances when fixed elements obscure the view of an element
that technical would be within the view port, but since the fixed
element is displayed over that element it is not clickable.

The fix for this was to bring the element as close to the center as
possible.